+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [*Unreleased*] - <*2019-04-29*>
+## [*1.2.0*] - <*2019-04-29*>
 
 ### Added
 * Added support for action command responses (`onFetchTask` - `composeExtension/fetchTask`)
+* Added support for `Action.Submit` from adaptive cards (`onCardButtonClicked` - `composeExtension/onCardButtonClicked`)
 
 ## [*1.1.0*] - <*2019-04-29*>